requirements.txt
PyQt5
mss
pyperclip
Pillow
pyzbar